requirements.txt
chromadb==1.0.7
Flask==3.1.0
flask-sqlalchemy==3.1.1
langchain==0.3.24
langchain-community==0.3.23
numpy==2.0.2
openai==1.76.2
pandas==2.2.3
pypdf==5.4.0
python-dotenv==1.1.0
tiktoken==0.9.0
